+++ extensions/voice/org.eclipse.smarthome.voice.mactts/build.properties
@@ -1,0 +1,5 @@
+source.. = src/main/java/
+output.. = target/classes
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/
