+++ extensions/voice/org.eclipse.smarthome.voice.mactts/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.eclipse.smarthome.voice</groupId>
+    <artifactId>pom</artifactId>
+    <version>0.8.0-SNAPSHOT</version>
+  </parent>
+
+  <groupId>org.eclipse.smarthome.voice</groupId>
+  <artifactId>org.eclipse.smarthome.voice.mactts</artifactId>
+  <version>0.8.0-SNAPSHOT</version>
+
+  <name>Eclipse SmartHome MacOS TTS</name>
+  <packaging>eclipse-plugin</packaging>
+
+</project>
