+++ extensions/voice/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="MACROMAN"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+  <parent>
+    <groupId>org.eclipse.smarthome.extension</groupId>
+    <artifactId>pom</artifactId>
+    <version>0.8.0-SNAPSHOT</version>
+  </parent>
+
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.eclipse.smarthome.voice</groupId>
+  <artifactId>pom</artifactId>
+
+  <name>Eclipse SmartHome Voice Services</name>
+
+  <packaging>pom</packaging>
+
+  <modules>
+    <module>org.eclipse.smarthome.voice.mactts</module>
+  </modules>
+
+</project>
